dev: added login failure handling with default display alert box

+++ flare-app/ViewModels/MainViewModel.cs
@@ -89,14 +89,19 @@ public partial class MainViewModel : ObservableObject
     /// </summary>
     async Task RemoveUser(string? s)
     {
+        if (s is null)
+            return;
         MyContact? removeThis = MyUsers.FirstOrDefault(u => u.ContactUserName == s && u.ContactOwner == "TempUser1");
-        try
+        if (removeThis is not null)
         {
-            await LocalUserDBService.DeleteContact(removeThis);
-            MyUsers.Remove(removeThis);
+            try
+            {
+                await LocalUserDBService.DeleteContact(removeThis);
+                MyUsers.Remove(removeThis);
+            }
+            catch { }
         }
-        catch { }
-        //await Refresh();
+        await Refresh();
     }
 
     /// <summary>
